Reference of config.xml fixed for android
If I used this plugin together with other plugins, the other plugins stopped to work on android. I checked the android's plugin.xml file and found that only this plugin's <feature>-Element was added.
This fix works for me, I don't know about side-effects, so use with caution - I just posted it because it may be helpful to others.
Conflicts:
	plugin.xml